% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstractArt.R
\name{make_art}
\alias{make_art}
\title{Random Art Function}
\usage{
make_art(
  seed = runif(1, 0, 1e+05),
  stars = sample(1:50, 1, replace = TRUE),
  starlines = sample(3:10, 1, replace = TRUE)
)
}
\arguments{
\item{seed}{Integer. Set a seed to make the picture reproducible. Defaults to a random value between 0 and 100000.}

\item{stars}{Integer. Defaults to a random value between 1 and 50.}

\item{starlines}{Integer. Defaults to a random value between 3 and 10.}
}
\description{
This function creates an image from randomly generated values.
More specifically, it creates a random amount of stars and starlines.
The colors are dependent on the amount of starlines.
}
\examples{
make_art(13580329)
}
\keyword{PiPS,}
\keyword{Q3.2R.2,}
\keyword{random_art}
